propagate extracted textgen-utils
