+++ packages/class-core-generator/package.json
@@ -30,6 +30,7 @@
     "@lionweb/core": "0.7.0-beta.1",
     "@lionweb/io-lionweb-mps-specific": "0.7.0-beta.1",
     "@lionweb/json": "0.7.0-beta.1",
+    "@lionweb/textgen-utils": "0.7.0-beta.1",
     "@lionweb/ts-utils": "0.7.0-beta.1",
     "@lionweb/utilities": "0.7.0-beta.1",
     "@types/node": "22.15.21",
