--- packages/class-core-generator/src/api/index.ts
@@ -1,20 +1,0 @@
-// Copyright 2025 TRUMPF Laser SE and other contributors
-//
-// Licensed under the Apache License, Version 2.0 (the "License")
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// SPDX-FileCopyrightText: 2025 TRUMPF Laser SE and other contributors
-// SPDX-License-Identifier: Apache-2.0
-
-export { generateLanguage, generateApiFromLanguages, generateApiFromLanguagesJson } from "./generator.js"
-export * from "./generation-headers.js"
-export * from "./mega-factory.templates.js"
