+++ packages/class-core-generator/src/generation-headers.ts
