+++ packages/class-core-generator/src/generator.ts
