+++ packages/class-core-generator/src/helpers/entities.ts
