+++ packages/class-core-generator/src/helpers/features.ts
