+++ packages/class-core-generator/src/helpers/index.ts
