+++ packages/class-core-generator/src/helpers/primitive-types.ts
