+++ packages/utilities/package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@lionweb/core": "0.7.0-beta.1",
     "@lionweb/json": "0.7.0-beta.1",
+    "@lionweb/textgen-utils": "0.7.0-beta.1",
     "@lionweb/ts-utils": "0.7.0-beta.1",
     "littoral-templates": "0.3.0",
     "nanoid": "5.0.9"
