--- packages/utilities/src/m3/ts-generation/textgen.ts
@@ -1,3 +1,0 @@
-import { indentWith } from "littoral-templates"
-
-export const indent = indentWith("    ")(1)
